requirements.txt
pymp-pypi
git+git://github.com/waspinator/pycococreator.git@0.2.0
ipdb~=0.13.3
yacs~=0.1.8
h5py~=2.10.0
Pillow~=8.0.1
opencv-python~=4.4.0.42
matplotlib~=3.3.1
cython
pycocotools~=2.0
plyfile~=0.7.2
scipy~=1.5.2
pandas~=1.1.1
seaborn~=0.10.1
scikit-image~=0.17.2
sklearn
tensorboardX~=2.1
hydra-core
setuptools~=50.3.2
numpy~=1.19.1
tqdm~=4.48.2
typing~=3.7.4.3
scikit-learn~=0.23.2
DateTime~=4.3
pytest~=6.0.1
imageio~=2.9.0
PyEXR~=0.3.9
tabulate~=0.8.7
easydict~=1.9
wandb
trimesh
open3d
hydra-colorlog
hydra-submitit-launcher
